requirements.txt
python-telegram-bot~=22.6
python-dotenv~=1.2.1
python-telegram-bot[job-queue]~=22.6